+++ packages/streaming-image-volume-loader/src/sharedArrayBufferImageLoader.ts
@@ -129,6 +129,8 @@ function getTransferSyntaxForContentType(contentType: string): string {
     'image/x-jls': '1.2.840.10008.1.2.4.80',
     'image/jp2': '1.2.840.10008.1.2.4.90',
     'image/jpx': '1.2.840.10008.1.2.4.92',
+    'image/jphc': '3.2.840.10008.1.2.4.96',
+    'image/jls': '1.2.840.10008.1.2.4.80',
   };
 
   if (params['transfer-syntax']) {