+++ packages/tools/src/tools/index.ts
@@ -20,6 +20,8 @@ import EllipticalROITool from './annotation/EllipticalROITool';
 import PlanarFreehandROITool from './annotation/PlanarFreehandROITool';
 import ArrowAnnotateTool from './annotation/ArrowAnnotateTool';
 import AngleTool from './annotation/AngleTool';
+import ReferenceCursors from './ReferenceCursors';
+import ReferenceLines from './ReferenceLinesTool';
 
 // Segmentation DisplayTool
 import SegmentationDisplayTool from './displayTools/SegmentationDisplayTool';
@@ -57,6 +59,7 @@ export {
   PlanarFreehandROITool,
   ArrowAnnotateTool,
   AngleTool,
+  ReferenceCursors,
   // Segmentations Display
   SegmentationDisplayTool,
   // Segmentations Tools
@@ -67,4 +70,5 @@ export {
   RectangleROIStartEndThresholdTool,
   BrushTool,
   MagnifyTool,
+  ReferenceLines,
 };