+++ packages/tools/src/utilities/segmentation/triggerSegmentationRender.ts
@@ -129,6 +129,10 @@ class SegmentationRenderingEngine {
     const segmentationDisplayToolInstance = toolGroup.getToolInstance(
       SegmentationDisplayTool.toolName
     ) as SegmentationDisplayTool;
+    if (!segmentationDisplayToolInstance) {
+      console.warn('No segmentation tool found inside', toolGroupId);
+      return;
+    }
 
     function onSegmentationRender(evt: Types.EventTypes.ImageRenderedEvent) {
       const { element, viewportId, renderingEngineId } = evt.detail;
@@ -139,6 +143,11 @@ class SegmentationRenderingEngine {
       );
 
       const toolGroup = getToolGroupForViewport(viewportId, renderingEngineId);
+
+      if (!toolGroup) {
+        console.warn('toolGroup has been destroyed');
+        return;
+      }
 
       const eventDetail: SegmentationRenderedEventDetail = {
         toolGroupId: toolGroup.id,